奖惩添加日志 列表展示填写人

+++ src/main/java/com/mwkj/lzda/service/impl/PunishServiceImpl.java
@@ -1,12 +1,17 @@
 package com.mwkj.lzda.service.impl;
 
+import com.mwkj.lzda.core.AppException;
 import com.mwkj.lzda.dao.PunishMapper;
 import com.mwkj.lzda.dto.PunishDTO;
+import com.mwkj.lzda.enu.ArchiveTypeEnum;
+import com.mwkj.lzda.enu.LogOperateTypeEnum;
+import com.mwkj.lzda.enu.PunishTypeEnum;
 import com.mwkj.lzda.model.*;
 import com.mwkj.lzda.service.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.ModelMap;
+import org.springframework.util.ObjectUtils;
 
 import javax.annotation.Resource;
 import java.util.List;
@@ -42,6 +47,9 @@ public class PunishServiceImpl implements PunishService {
 
     @Resource
     PunGiftService punGiftService;
+
+    @Resource
+    OperateLogService operateLogService;
 
     /**
      * @return java.util.List<com.mwkj.lzda.dto.PunishDTO>
@@ -103,6 +111,7 @@ public class PunishServiceImpl implements PunishService {
     @Override
     public String toPunishDetail(PunishDTO punishDTO, ModelMap map) {
         String page = null;
+        int userId = 0;
         switch (punishDTO.getPunishType()) {
             case 1:
                 PunViolation punViolation = punViolationService.findById(punishDTO.getPunishId());
@@ -110,6 +119,7 @@ public class PunishServiceImpl implements PunishService {
                 map.put("punish", punViolation);
                 map.put("attachments", attachments1);
                 page = "/views/punish/pun_violation_table";
+                userId = punViolation.getUserId();
                 break;
             case 2:
                 PunAccountability punAccountability = punAccountabilityService.findById(punishDTO.getPunishId());
@@ -117,6 +127,7 @@ public class PunishServiceImpl implements PunishService {
                 map.put("punish", punAccountability);
                 map.put("attachments", attachments2);
                 page = "/views/punish/pun_accountability_table";
+                userId = punAccountability.getUserId();
                 break;
             case 3:
                 PunGift punGift = punGiftService.findById(punishDTO.getPunishId());
@@ -124,6 +135,7 @@ public class PunishServiceImpl implements PunishService {
                 map.put("punish", punGift);
                 map.put("attachments", attachments3);
                 page = "/views/punish/pun_gift_table";
+                userId = punGift.getUserId();
                 break;
             case 4:
                 PunTalk punTalk = punTalkService.findById(punishDTO.getPunishId());
@@ -131,6 +143,7 @@ public class PunishServiceImpl implements PunishService {
                 map.put("punish", punTalk);
                 map.put("attachments", attachments4);
                 page = "/views/punish/pun_talk_table";
+                userId = punTalk.getUserId();
                 break;
             case 5:
                 PunNotice punNotice = punNoticeService.findById(punishDTO.getPunishId());
@@ -138,6 +151,7 @@ public class PunishServiceImpl implements PunishService {
                 map.put("punish", punNotice);
                 map.put("attachments", attachments5);
                 page = "/views/punish/pun_notice_table";
+                userId = punNotice.getUserId();
                 break;
             case 6:
                 PunReport punReport = punReportService.findById(punishDTO.getPunishId());
@@ -145,31 +159,91 @@ public class PunishServiceImpl implements PunishService {
                 map.put("punish", punReport);
                 map.put("attachments", attachments6);
                 page = "/views/punish/pun_report_table";
+                userId = punReport.getUserId();
                 break;
         }
+
+        //遍历惩罚表类型枚举
+        String operateObject = null;
+        for (PunishTypeEnum typeEnum : PunishTypeEnum.values()) {
+            if (punishDTO.getPunishType() == typeEnum.punishType()) {
+                operateObject = typeEnum.punishName();
+                break;
+            }
+        }
+        //插入日志信息
+        operateLogService.save(operateObject, LogOperateTypeEnum.查看.toString(), userId);
         return page;
     }
 
     @Override
     public void delete(PunishDTO punishDTO) {
+
+        //遍历惩罚表类型枚举
+        String operateObject = null;
+        for (PunishTypeEnum typeEnum : PunishTypeEnum.values()) {
+            if (punishDTO.getPunishType() == typeEnum.punishType()) {
+                operateObject = typeEnum.punishName();
+                break;
+            }
+        }
+
         switch (punishDTO.getPunishType()) {
             case 1:
+                PunViolation punViolation = punViolationService.findById(punishDTO.getPunishId());
+                if (ObjectUtils.isEmpty(punViolation)) {
+                    throw new AppException("删除失败，数据不存在");
+                }
                 punViolationService.deleteById(punishDTO.getPunishId());
+
+                //插入日志信息
+                operateLogService.save(operateObject, LogOperateTypeEnum.删除.toString(), punViolation.getUserId());
                 break;
             case 2:
+                PunAccountability punAccountability = punAccountabilityService.findById(punishDTO.getPunishId());
+                if (ObjectUtils.isEmpty(punAccountability)) {
+                    throw new AppException("删除失败，数据不存在");
+                }
                 punAccountabilityService.deleteById(punishDTO.getPunishId());
+
+                //插入日志信息
+                operateLogService.save(operateObject, LogOperateTypeEnum.删除.toString(), punAccountability.getUserId());
                 break;
             case 3:
+                PunGift punGift = punGiftService.findById(punishDTO.getPunishId());
+                if (ObjectUtils.isEmpty(punGift)) {
+                    throw new AppException("删除失败，数据不存在");
+                }
                 punGiftService.deleteById(punishDTO.getPunishId());
+                //插入日志信息
+                operateLogService.save(operateObject, LogOperateTypeEnum.删除.toString(), punGift.getUserId());
                 break;
             case 4:
+                PunTalk punTalk = punTalkService.findById(punishDTO.getPunishId());
+                if (ObjectUtils.isEmpty(punTalk)) {
+                    throw new AppException("删除失败，数据不存在");
+                }
                 punTalkService.deleteById(punishDTO.getPunishId());
+                //插入日志信息
+                operateLogService.save(operateObject, LogOperateTypeEnum.删除.toString(), punTalk.getUserId());
                 break;
             case 5:
+                PunNotice punNotice = punNoticeService.findById(punishDTO.getPunishId());
+                if (ObjectUtils.isEmpty(punNotice)) {
+                    throw new AppException("删除失败，数据不存在");
+                }
                 punNoticeService.deleteById(punishDTO.getPunishId());
+                //插入日志信息
+                operateLogService.save(operateObject, LogOperateTypeEnum.删除.toString(), punNotice.getUserId());
                 break;
             case 6:
+                PunReport punReport = punReportService.findById(punishDTO.getPunishId());
+                if (ObjectUtils.isEmpty(punReport)) {
+                    throw new AppException("删除失败，数据不存在");
+                }
                 punReportService.deleteById(punishDTO.getPunishId());
+                //插入日志信息
+                operateLogService.save(operateObject, LogOperateTypeEnum.删除.toString(), punReport.getUserId());
                 break;
         }
     }
